move wallets.txt parsing logic to casper.util.VaultsParser
